Extract functional utility template functions into a functional Jinja2 extension
