+++ homeassistant/helpers/template/__init__.py
@@ -6,15 +6,12 @@ from ast import literal_eval
 import asyncio
 import collections.abc
 from collections.abc import Callable, Generator, Iterable
-from copy import deepcopy
 from datetime import datetime, timedelta
 from enum import Enum
 from functools import cache, lru_cache, partial, wraps
 import logging
 import math
-from operator import contains
 import pathlib
-import random
 import re
 import sys
 from types import CodeType
@@ -45,7 +42,6 @@ from homeassistant.const import (
 from homeassistant.core import (
     Context,
     HomeAssistant,
-    ServiceResponse,
     State,
     callback,
     valid_domain,
@@ -1460,162 +1456,9 @@ def add(value, amount, default=_SENTINEL):
         return default
 
 
-def apply(value, fn, *args, **kwargs):
-    """Call the given callable with the provided arguments and keyword arguments."""
-    return fn(value, *args, **kwargs)
-
-
-def as_function(macro: jinja2.runtime.Macro) -> Callable[..., Any]:
-    """Turn a macro with a 'returns' keyword argument into a function that returns what that argument is called with."""
-
-    def wrapper(*args, **kwargs):
-        return_value = None
-
-        def returns(value):
-            nonlocal return_value
-            return_value = value
-            return value
-
-        # Call the callable with the value and other args
-        macro(*args, **kwargs, returns=returns)
-        return return_value
-
-    # Remove "macro_" from the macro's name to avoid confusion in the wrapper's name
-    trimmed_name = macro.name.removeprefix("macro_")
-
-    wrapper.__name__ = trimmed_name
-    wrapper.__qualname__ = trimmed_name
-    return wrapper
-
-
 def version(value):
     """Filter and function to get version object of the value."""
     return AwesomeVersion(value)
-
-
-def merge_response(value: ServiceResponse) -> list[Any]:
-    """Merge action responses into single list.
-
-    Checks that the input is a correct service response:
-    {
-        "entity_id": {str: dict[str, Any]},
-    }
-    If response is a single list, it will extend the list with the items
-        and add the entity_id and value_key to each dictionary for reference.
-    If response is a dictionary or multiple lists,
-        it will append the dictionary/lists to the list
-        and add the entity_id to each dictionary for reference.
-    """
-    if not isinstance(value, dict):
-        raise TypeError("Response is not a dictionary")
-    if not value:
-        # Bail out early if response is an empty dictionary
-        return []
-
-    is_single_list = False
-    response_items: list = []
-    input_service_response = deepcopy(value)
-    for entity_id, entity_response in input_service_response.items():  # pylint: disable=too-many-nested-blocks
-        if not isinstance(entity_response, dict):
-            raise TypeError("Response is not a dictionary")
-        for value_key, type_response in entity_response.items():
-            if len(entity_response) == 1 and isinstance(type_response, list):
-                # Provides special handling for responses such as calendar events
-                # and weather forecasts where the response contains a single list with multiple
-                # dictionaries inside.
-                is_single_list = True
-                for dict_in_list in type_response:
-                    if isinstance(dict_in_list, dict):
-                        if ATTR_ENTITY_ID in dict_in_list:
-                            raise ValueError(
-                                f"Response dictionary already contains key '{ATTR_ENTITY_ID}'"
-                            )
-                        dict_in_list[ATTR_ENTITY_ID] = entity_id
-                        dict_in_list["value_key"] = value_key
-                response_items.extend(type_response)
-            else:
-                # Break the loop if not a single list as the logic is then managed in the outer loop
-                # which handles both dictionaries and in the case of multiple lists.
-                break
-
-        if not is_single_list:
-            _response = entity_response.copy()
-            if ATTR_ENTITY_ID in _response:
-                raise ValueError(
-                    f"Response dictionary already contains key '{ATTR_ENTITY_ID}'"
-                )
-            _response[ATTR_ENTITY_ID] = entity_id
-            response_items.append(_response)
-
-    return response_items
-
-
-def fail_when_undefined(value):
-    """Filter to force a failure when the value is undefined."""
-    if isinstance(value, jinja2.Undefined):
-        value()
-    return value
-
-
-@pass_context
-def random_every_time(context, values):
-    """Choose a random value.
-
-    Unlike Jinja's random filter,
-    this is context-dependent to avoid caching the chosen value.
-    """
-    return random.choice(values)
-
-
-def iif(
-    value: Any, if_true: Any = True, if_false: Any = False, if_none: Any = _SENTINEL
-) -> Any:
-    """Immediate if function/filter that allow for common if/else constructs.
-
-    https://en.wikipedia.org/wiki/IIf
-
-    Examples:
-        {{ is_state("device_tracker.frenck", "home") | iif("yes", "no") }}
-        {{ iif(1==2, "yes", "no") }}
-        {{ (1 == 1) | iif("yes", "no") }}
-
-    """
-    if value is None and if_none is not _SENTINEL:
-        return if_none
-    if bool(value):
-        return if_true
-    return if_false
-
-
-def typeof(value: Any) -> Any:
-    """Return the type of value passed to debug types."""
-    return value.__class__.__name__
-
-
-def combine(*args: Any, recursive: bool = False) -> dict[Any, Any]:
-    """Combine multiple dictionaries into one."""
-    if not args:
-        raise TypeError("combine expected at least 1 argument, got 0")
-
-    result: dict[Any, Any] = {}
-    for arg in args:
-        if not isinstance(arg, dict):
-            raise TypeError(f"combine expected a dict, got {type(arg).__name__}")
-
-        if recursive:
-            for key, value in arg.items():
-                if (
-                    key in result
-                    and isinstance(result[key], dict)
-                    and isinstance(value, dict)
-                ):
-                    result[key] = combine(result[key], value, recursive=True)
-                else:
-                    result[key] = value
-        else:
-            result |= arg
-
-    return result
 
 
 def make_logging_undefined(
@@ -1754,6 +1597,9 @@ class TemplateEnvironment(ImmutableSandboxedEnvironment):
         )
         self.add_extension("homeassistant.helpers.template.extensions.DeviceExtension")
         self.add_extension("homeassistant.helpers.template.extensions.FloorExtension")
+        self.add_extension(
+            "homeassistant.helpers.template.extensions.FunctionalExtension"
+        )
         self.add_extension("homeassistant.helpers.template.extensions.IssuesExtension")
         self.add_extension("homeassistant.helpers.template.extensions.LabelExtension")
         self.add_extension("homeassistant.helpers.template.extensions.MathExtension")
@@ -1766,31 +1612,12 @@ class TemplateEnvironment(ImmutableSandboxedEnvironment):
             "homeassistant.helpers.template.extensions.TypeCastExtension"
         )
 
-        self.globals["apply"] = apply
-        self.globals["as_function"] = as_function
-        self.globals["combine"] = combine
-        self.globals["iif"] = iif
-        self.globals["merge_response"] = merge_response
-        self.globals["typeof"] = typeof
         self.globals["version"] = version
-        self.globals["zip"] = zip
 
         self.filters["add"] = add
-        self.filters["apply"] = apply
-        self.filters["as_function"] = as_function
-        self.filters["combine"] = combine
-        self.filters["contains"] = contains
-        self.filters["iif"] = iif
-        self.filters["is_defined"] = fail_when_undefined
         self.filters["multiply"] = multiply
-        self.filters["ord"] = ord
-        self.filters["random"] = random_every_time
         self.filters["round"] = forgiving_round
-        self.filters["typeof"] = typeof
         self.filters["version"] = version
-
-        self.tests["apply"] = apply
-        self.tests["contains"] = contains
 
         if hass is None:
             return
